Refactor: Add index.js in models, controllers, services for importing and exporting
