+++ src/controllers/auth.controller.js
@@ -1,6 +1,5 @@
-const authService = require('../services/auth.services.js');
-const AppError = require('../utils/AppError.js');
-const userService = require('../services/user.services.js');
+const { userService, authService } = require('../services');
+const AppError = require('../utils/AppError');
 
 const createTokenAndSend = (user, res) => {
   const token = authService.generateAuthToken(user._id);
@@ -10,7 +9,6 @@ const createTokenAndSend = (user, res) => {
     user: user
   });
 };
-
 
 /**
  * @version 1.0.0
@@ -27,11 +25,10 @@ exports.signup = async (req, res, next) => {
   }
   const newUser = await userService.createUser(req.body);
   // TODO
-  // Return 401 if role is premium without credit 
+  // Return 401 if role is premium without credit
   // Return 401 if role is artist without request
   // Add device
   // use mail to verify user
-
   createTokenAndSend(newUser, res);
 };
 
@@ -44,12 +41,12 @@ exports.signup = async (req, res, next) => {
  * @summary User Login
  */
 exports.login = async (req, res, next) => {
-  const {
-    email,
-    password
-  } = req.body;
+  const { email, password } = req.body;
 
-  const user = await userService.findUserAndCheckPassword({ email: email }, password);
+  const user = await userService.findUserAndCheckPassword(
+    { email: email },
+    password
+  );
   if (!user) {
     return next(new AppError('Incorrect email or password!', 401));
   }
@@ -64,18 +61,14 @@ exports.login = async (req, res, next) => {
 
 /**
  * @version 1.0.0
- * @throws AppError 400 status, AppError 401 status 
+ * @throws AppError 400 status, AppError 401 status
  * @author Abdelrahman Tarek
  * @description takes currentPassword, Password and passwordConfirm. if currentPassword is wrong return 401 status
  * if password != passwordConfirm return 400
  * @summary User Update Password
  */
 exports.updatePassword = async (req, res, next) => {
-  const {
-    currentPassword,
-    password,
-    passwordConfirm
-  } = req.body;
+  const { currentPassword, password, passwordConfirm } = req.body;
 
   if (!req.user) {
     return next(new AppError('PLease Authentcate first', 500));
@@ -85,14 +78,16 @@ exports.updatePassword = async (req, res, next) => {
     return next(new AppError('Please confirm your password', 400));
   }
 
-  const user = await userService.findUserByIdAndCheckPassword(req.user._id, currentPassword);
+  const user = await userService.findUserByIdAndCheckPassword(
+    req.user._id,
+    currentPassword
+  );
   if (!user) {
     return next(new AppError('Incorrect password!', 401));
   }
 
   user.password = password;
   user.passwordConfirm = password;
-
 
   // TODO
   // Add last change password date
