+++ src/controllers/index.js
@@ -1,0 +1,2 @@
+exports.authController = require('./auth.controller');
+exports.homeController = require('./home.controller');
