+++ src/models/album.model.js
@@ -49,4 +49,4 @@ const albumSchema = new mongoose.Schema({
 
 const Album = mongoose.model('Album', albumSchema);
 
-module.exports = Album;
+module.exports = { Album, albumSchema };
