+++ src/models/artist.model.js
@@ -31,4 +31,4 @@ const artistSchema = new mongoose.Schema({
 
 const Artist = mongoose.model('Artist', artistSchema);
 
-module.exports = Artist;
+module.exports = { Artist, artistSchema };
