+++ src/models/index.js
@@ -1,0 +1,15 @@
+const { User, userSchema } = require('./user.model');
+const { Artist, artistSchema } = require('./artist.model');
+const { Track, trackSchema } = require('./track.model');
+const { Album, albumSchema } = require('./album.model');
+
+module.exports = {
+  User,
+  userSchema,
+  Artist,
+  artistSchema,
+  Track,
+  trackSchema,
+  Album,
+  albumSchema
+};
