+++ src/models/track.model.js
@@ -33,4 +33,4 @@ const trackSchema = new mongoose.Schema({
 
 const Track = mongoose.model('Track', trackSchema);
 
-module.exports = Track;
+module.exports = { Track, trackSchema };
