+++ src/routes/v1/user.route.js
@@ -1,8 +1,8 @@
 const express = require('express');
-const authController = require('../../controllers/auth.controller.js');
-const validate = require('../../middlewares/validate.js');
-const authValidation = require('../../validations/auth.validation.js');
-const catchAsync = require('../../utils/catchAsync.js');
+const { authController } = require('../../controllers');
+const validate = require('../../middlewares/validate');
+const { authValidation } = require('../../validations');
+const catchAsync = require('../../utils/catchAsync');
 
 const router = express.Router();
