+++ src/services/index.js
@@ -1,0 +1,2 @@
+exports.authService = require('./auth.service');
+exports.userService = require('./user.service');
