+++ src/services/user.service.js
@@ -1,40 +1,40 @@
-const { User } = require('../models/user.model.js');
-const authService = require('./auth.services.js');
+const { User } = require('../models');
+const authService = require('./auth.service');
 
 /**
- * 
- * @param {Object} userData 
- * @param {String} password 
+ *
+ * @param {Object} userData
+ * @param {String} password
  * @returns user if user found and password is correct else return null
  */
 const findUserAndCheckPassword = async (userData, password) => {
   const user = await User.findOne(userData).select('+password');
-  if (!user || !await authService.checkPassword(password, user.password)) {
+  if (!user || !(await authService.checkPassword(password, user.password))) {
     return null;
   }
   return user;
 };
 
 /**
- * 
- * @param {String} userId 
- * @param {String} password 
+ *
+ * @param {String} userId
+ * @param {String} password
  * @returns user if user found and password is correct else return null
  */
 const findUserByIdAndCheckPassword = async (userId, password) => {
   const user = await User.findById(userId).select('+password');
-  if (!user || !await authService.checkPassword(password, user.password)) {
+  if (!user || !(await authService.checkPassword(password, user.password))) {
     return null;
   }
   return user;
 };
 
 /**
- * 
- * @param {Object} userData 
+ *
+ * @param {Object} userData
  * @returns new user
  */
-const createUser = async (userData) => {
+const createUser = async userData => {
   const newUser = await User.create(userData);
   return newUser;
 };
@@ -43,4 +43,4 @@ module.exports = {
   findUserAndCheckPassword,
   findUserByIdAndCheckPassword,
   createUser
-}
+};
