+++ src/validations/index.js
@@ -1,0 +1,1 @@
+exports.authValidation = require('./auth.validation');
